Misc 1.5.0 improvements (#69)
* tsconfig.build.json rename

* clenaed up header provider resolution in vectorize tests

* Updated

* Automate Accounting

* updated error messages

* added test devop url

* repl script

* Don’t even try to refactor it.

* added astra cli to shell.nix

* minor doc fix

* fixed mistaken bug

* updated typescript + added "-light" build option

* pagingState => pageState

* fixed some failing tests

* updated deps, fixed test, disabled returnDocumentResponses

* I don't get paid enough for this shit.

* improved some test coverage

* I just wanna tell you how I'm feeling

* udpated typescript

* done. going to bed now.

* fixed devops lifecycle test

* added more valid db statuses

+++ scripts/build.sh
@@ -7,27 +7,33 @@ rm -rf ./dist
 node scripts/utils/build-version-file.js > src/version.ts
 
 # Transpiles the project
-npx tsc --project tsconfig.build.json
+if [ "$1" = "-light" ]; then
+  npx tsc --project tsconfig.production.json -d false --noCheck
+else
+  npx tsc --project tsconfig.production.json
+fi
 
 # Replaces alias paths with relative paths (e.g. `@/src/version` -> `../../src/version`)
-npx tsc-alias -p tsconfig.build.json
+npx tsc-alias -p tsconfig.production.json
 
-# Creates the rollup .d.ts, generates an API report in etc/, and cleans up any temp files
-npx api-extractor run -c ./api-extractor.jsonc --local && rm -r ./temp
+if [ "$1" != "-light" ]; then
+  # Creates the rollup .d.ts, generates an API report in etc/, and cleans up any temp files
+  npx api-extractor run -c ./api-extractor.jsonc --local && rm -r ./temp
 
-# Uses a more succinct licence notice + removes block comments (the rollup .d.ts file already contains the ts-doc)
-find ./dist -type f -name '*.js' -exec node scripts/utils/reduce-comments.js {} \;
+  # Uses a more succinct licence notice + removes block comments (the rollup .d.ts file already contains the ts-doc)
+  find ./dist -type f -name '*.js' -exec node scripts/utils/reduce-comments.js {} \;
 
-# Adds the missing license notice to the rollup .d.ts
-node scripts/utils/add-license-bumf.js dist/astra-db-ts.d.ts
+  # Adds the missing license notice to the rollup .d.ts
+  node scripts/utils/add-license-bumf.js dist/astra-db-ts.d.ts
 
-# Delete the "empty" files where only types were declared
-node scripts/utils/del-empty-dist-files.js
+  # Delete the "empty" files where only types were declared
+  node scripts/utils/del-empty-dist-files.js
 
-# Removes all .d.ts files except the main rollup .d.ts
-cd dist || return 1
-find . -type f -name '*.d.ts' ! -name 'astra-db-ts.d.ts' -exec rm {} +
-cd ..
+  # Removes all .d.ts files except the main rollup .d.ts
+  cd dist || return 1
+  find . -type f -name '*.d.ts' ! -name 'astra-db-ts.d.ts' -exec rm {} +
+  cd ..
 
-# Delete any empty leftover directories
-find ./dist -type d -empty -delete
+  # Delete any empty leftover directories
+  find ./dist -type d -empty -delete
+fi
